+++ src/ec_sys_linux.h
@@ -9,9 +9,9 @@ extern EC_VTable EC_SysLinux_ACPI_VTable;
 Error* EC_SysLinux_Open();
 Error* EC_SysLinux_ACPI_Open();
 void   EC_SysLinux_Close();
-Error* EC_SysLinux_WriteByte(int, uint8_t);
-Error* EC_SysLinux_WriteWord(int, uint16_t);
-Error* EC_SysLinux_ReadByte(int, uint8_t*);
-Error* EC_SysLinux_ReadWord(int, uint16_t*);
+Error* EC_SysLinux_WriteByte(uint8_t, uint8_t);
+Error* EC_SysLinux_WriteWord(uint8_t, uint16_t);
+Error* EC_SysLinux_ReadByte(uint8_t, uint8_t*);
+Error* EC_SysLinux_ReadWord(uint8_t, uint16_t*);
 
 #endif
